#3 Improve CI, generate and build the artifact
Execute example test during integration test phase of archetype generation.
