+++ src/main/resources/archetype-resources/pom.xml
@@ -47,6 +47,19 @@
 					<target>${java.version}</target>
 				</configuration>
 			</plugin>
+			
+			<plugin>
+        <groupId>org.apache.maven.plugins</groupId>
+        <artifactId>maven-surefire-plugin</artifactId>
+        <version>2.22.2</version>
+        <configuration>
+          <properties>
+            <configurationParameters>
+              junit.jupiter.extensions.autodetection.enabled=true
+            </configurationParameters>
+          </properties>
+        </configuration>
+      </plugin>
 		</plugins>
 	</build>
 </project>
