+++ src/test/resources/projects/it-basic/archetype.properties
@@ -1,4 +1,4 @@
-groupId=de.retest.archetypes
-artifactId=recheck-web-archetype
+groupId=de.retest.example
+artifactId=recheck-web-example
 version=1.0-SNAPSHOT
-package=it.pkg
+package=recheck.example
